app/requirements.txt
# Server
# ---------------------------------------------
Django==2.0.6
django_rq==1.2.0
gunicorn==19.8.1
whitenoise==3.3.1

# Caching
# ---------------------------------------------
python3-memcached

# General
# ---------------------------------------------
lti==0.9.3
oauthlib==2.1.0
python-dotenv==0.8.2
requests==2.18.4

# Database
# ---------------------------------------------
psycopg2==2.7.5

# NLP
# ---------------------------------------------
git+https://github.com/hltcoe/PredPatt.git
flashtext==2.7
gensim==3.5.0
nltk==3.3
pyspellchecker==0.1.2
sacremoses==0.0.3

# Testing
# ---------------------------------------------
coverage-badge==0.2.0
pytest==3.6.2
pytest-cov==2.5.1
